messing around searching for cleaness
