+++ bioptim/optimization/non_linear_program.py
@@ -2,11 +2,12 @@ from typing import Callable, Any
 
 import casadi
 from casadi import SX, MX, Function, horzcat
+import numpy as np
 
 from .optimization_variable import OptimizationVariable, OptimizationVariableContainer
 from ..dynamics.ode_solver import OdeSolver
 from ..limits.path_conditions import InitialGuessList, BoundsList
-from ..misc.enums import ControlType, PhaseDynamics
+from ..misc.enums import ControlType, PhaseDynamics, Shooting
 from ..misc.options import OptionList
 from ..misc.mapping import NodeMapping
 from ..dynamics.dynamics_evaluation import DynamicsEvaluation
@@ -419,3 +420,104 @@ class NonLinearProgram:
         if node_idx < 0 or node_idx > self.ns:
             return ValueError(f"node_index out of range [0:{self.ns}]")
         return self.tf / self.ns * node_idx
+
+    def _generate_time(self,
+                       keep_intermediate_points: bool = None,
+                       shooting_type: Shooting = None,
+                       ):
+
+        is_direct_collocation = self.ode_solver.is_direct_collocation
+        include_starting_collocation_point = False
+        if is_direct_collocation:
+            include_starting_collocation_point = self.ode_solver.include_starting_collocation_point
+
+        step_times = self._define_step_times(
+            dynamics_step_time=self.dynamics[0].step_time,
+            ode_solver_steps=self.ode_solver.steps,
+            is_direct_collocation=is_direct_collocation,
+            include_starting_collocation_point=include_starting_collocation_point,
+            keep_intermediate_points=keep_intermediate_points,
+            continuous=shooting_type == Shooting.SINGLE,
+        )
+
+        if shooting_type == Shooting.SINGLE_DISCONTINUOUS_PHASE:
+            # discard the last time step because continuity concerns only the end of the phases
+            # and not the end of each interval
+            step_times = step_times[:-1]
+
+        dt_ns = time_phase[p + 1] / self.ns
+        time = [(step_times * dt_ns + i * dt_ns).tolist() for i in range(nlp.ns)]
+
+        if shooting_type == Shooting.MULTIPLE:
+            # keep all the intervals in separate lists
+            flat_time = [np.array(sub_time) for sub_time in time]
+        else:
+            # flatten the list of list into a list of floats
+            flat_time = [st for sub_time in time for st in sub_time]
+
+        # add the final time of the phase
+        if shooting_type == Shooting.MULTIPLE:
+            flat_time.append(np.array([nlp.ns * dt_ns]))
+        if shooting_type == Shooting.SINGLE or shooting_type == Shooting.SINGLE_DISCONTINUOUS_PHASE:
+            flat_time += [nlp.ns * dt_ns]
+
+        return sum(time_phase[: p + 1]) + np.array(flat_time, dtype=object)
+
+    @staticmethod
+    def _define_step_times(
+            dynamics_step_time: list,
+            ode_solver_steps: int,
+            keep_intermediate_points: bool = None,
+            continuous: bool = True,
+            is_direct_collocation: bool = None,
+            include_starting_collocation_point: bool = False,
+    ) -> np.ndarray:
+        """
+        Define the time steps for the integration of the whole phase
+
+        Parameters
+        ----------
+        dynamics_step_time: list
+            The step time of the dynamics function
+        ode_solver_steps: int
+            The number of steps of the ode solver
+        keep_intermediate_points: bool
+            If the integration should return the intermediate values of the integration [False]
+            or only keep the node [True] effective keeping the initial size of the states
+        continuous: bool
+            If the arrival value of a node should be discarded [True] or kept [False]. The value of an integrated
+            arrival node and the beginning of the next one are expected to be almost equal when the problem converged
+        is_direct_collocation: bool
+            If the ode solver is direct collocation
+        include_starting_collocation_point
+            If the ode solver is direct collocation and an additional collocation point at the shooting node was used
+
+        Returns
+        -------
+        step_times: np.ndarray
+            The time steps for each interval of the phase of ocp
+
+        # TODO: move it somewhere else too :)
+        """
+
+        if keep_intermediate_points is None:
+            keep_intermediate_points = True if is_direct_collocation else False
+
+        if is_direct_collocation:
+            # time is not linear because of the collocation points
+            if keep_intermediate_points:
+                step_times = np.array(dynamics_step_time + [1])
+
+                if include_starting_collocation_point:
+                    step_times = np.array([0] + step_times)
+            else:
+                step_times = np.array(dynamics_step_time + [1])[[0, -1]]
+
+        else:
+            # time is linear in the case of direct multiple shooting
+            step_times = np.linspace(0, 1, ode_solver_steps + 1) if keep_intermediate_points else np.array([0, 1])
+        # it does not take the last nodes of each interval
+        if continuous:
+            step_times = step_times[:-1]
+
+        return step_times
